requirements (2).txt
torch
transformers
sentence-transformers
gradio
faiss-cpu
sacrebleu
matplotlib
